+++ src/App.tsx
@@ -1,14 +1,12 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import "./App.css";
 import HomePage from "./components/HomePage";
-import CreateProduct from "./components/CreateProduct";
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<HomePage />}></Route>
-        <Route path="/create" element={<CreateProduct />}></Route>
       </Routes>
     </BrowserRouter>
   );
