+++ src/components/CreateProduct.tsx
@@ -1,8 +1,4 @@
 import { useState } from "react";
-import Button from "@mui/material/Button";
-import Snackbar, { SnackbarCloseReason } from "@mui/material/Snackbar";
-import IconButton from "@mui/material/IconButton";
-import CloseIcon from "@mui/icons-material/Close";
 import "../assets/css/createproduct.css";
 import axios from "axios";
 import React from "react";
@@ -18,50 +14,16 @@ export default function CreateProduct({
     imageUrl: "",
   });
 
-  const [open, setOpen] = useState(false);
-
-  const handleSubmit = (
-    event:
-      | React.FormEvent<HTMLFormElement>
-      | React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
     axios
       .post("http://localhost:3000/products", product)
       .then((res) => {
         console.log(res);
         onProductCreated();
-        setOpen(true);
       })
       .catch((err) => console.log(err));
   };
-
-  const handleClose = (
-    _event: React.SyntheticEvent | Event,
-    reason?: SnackbarCloseReason
-  ) => {
-    if (reason === "clickaway") {
-      return;
-    }
-
-    setOpen(false);
-  };
-
-  const action = (
-    <React.Fragment>
-      <Button color="secondary" size="small" onClick={handleClose}>
-        UNDO
-      </Button>
-      <IconButton
-        size="small"
-        aria-label="close"
-        color="inherit"
-        onClick={handleClose}
-      >
-        <CloseIcon fontSize="small" />
-      </IconButton>
-    </React.Fragment>
-  );
 
   return (
     <div>
@@ -106,15 +68,6 @@ export default function CreateProduct({
           <button type="submit">Add Product</button>
         </div>
       </form>
-      <div>
-        <Snackbar
-          open={open}
-          autoHideDuration={6000}
-          onClose={handleClose}
-          message="Note archived"
-          action={action}
-        />
-      </div>
     </div>
   );
 }
